Added 7th class and fixed dataset split
Added fist gesture from each subject to the data set.

Made the split between training, validation, and testing subsets
of data. Testing subset is only to be used on fully trained model.
Writing info about number of samples for each class that are included in
 each subset to txt file.

Added the extension .npy hard coded in the class initializtion so it
should not be in the path passed to create the class object so the path
variable can be reused.