Modified saving of weights; Added commandline args
Saving history of weights in list that lines up with loss history so the
lowest cost can be used to select the weights. Reduces concern that best
weights are being saved with current if statement. Also useful for
saving weight history.

Added command line optional arguments to train_net to make choosing to
save models or just run test case is easier.