Script for PNN modified to except N subjects
pnn_train, a modified version of train_net, is able to read in N number
of subjects that will each have a model to be train, that will then be
added to the source model. Modifications were made to all dictionaries
that only needed one subject in the train_net context (data_loaders,
histories of loss,accuracy, and weights, etc).

For each epoch, training function loops through each subject to train it
with its data loader, then adds the layer weights (model.parameters())
to the source model to create a new model.